Validations

+++ app/Http/Controllers/MarcaController.php
@@ -19,7 +19,7 @@ class MarcaController extends Controller
     public function index()
     {
         $marcas = $this->marca->all();
-        return $marcas;
+        return response()->json($marcas, 200);
     }
 
     /**
@@ -30,8 +30,11 @@ class MarcaController extends Controller
      */
     public function store(Request $request)
     {
+
+        $request->validate($this->marca->rules(), $this->marca->feedback());
+
         $marca = $this->marca->create($request->all());
-        return $marca;
+        return response()->json($marca, 201);
     }
 
     /**
@@ -43,7 +46,10 @@ class MarcaController extends Controller
     public function show($id)
     {
         $marca = $this->marca->find($id);
-        return $marca;
+        if ($marca === null) {
+            return response()->json(['erro' => 'O registro solicitado não existe'], 404);
+        }
+        return response()->json($marca, 200);
     }
 
     /**
@@ -56,8 +62,28 @@ class MarcaController extends Controller
     public function update(Request $request, $id)
     {
         $marca = $this->marca->find($id);
+
+        if ($marca === null) {
+            return response()->json(['erro' => 'Não foi possível realizar a atualização. Registro inexistente'], 404);
+        }
+
+        if ($request->method() === 'PATCH') {
+            $regrasDinamicas = array();
+
+            foreach($marca->rules() as $input => $regra) {
+                if (array_key_exists($input, $request->all())) {
+                    $regrasDinamicas[$input] = $regra;
+                }
+            }
+
+            $request->validate($regrasDinamicas, $marca->feedback());
+        } else {
+            $request->validate($marca->rules(), $marca->feedback());
+        }
+
+
         $marca->update($request->all());
-        return $marca;
+        return response()->json($marca, 200);
     }
 
     /**
@@ -69,7 +95,12 @@ class MarcaController extends Controller
     public function destroy($id)
     {
         $marca = $this->marca->find($id);
+
+        if ($marca === null) {
+            return response()->json(['erro' => 'Não foi possível realizar a exclusão. Registro inexistente'], 404);
+        }
+
         $marca->delete();
-        return ['msg' => 'A marca foi excluída com sucesso'];
+        return response()->json(['msg' => 'A marca foi excluída com sucesso'], 200);
     }
 }
